+++ tests/view/newPerspectives/utils.js
@@ -7,7 +7,7 @@
  */
 
 /**
- * tests/view/components/perspectiveStatic.js
+ * tests/view/perspectives/utils.js
  */
 
 import { expect } from 'chai';
@@ -16,8 +16,37 @@ import { getArray,
   getConfig,
   filteredArray,
   getOptions,
-} from '../../../view/perspective/utils';
-import { getSubjects } from './utils';
+} from '../../../view/newPerspective/utils';
+
+const ZERO = 0;
+
+/**
+ * Returns an array of resources with identical
+ * isPublished property, with
+ * fieldName field == index in loop
+ *
+ * @param {Integer} INT Make this many resources
+ * @param {String} fieldName The field of each resource
+ * @param {Boolean} isPublished All resources have
+ * this value of isPublished
+ * @returns {Array} Array with all published resources
+ */
+function getSubjects(INT, fieldName, isPublished) {
+  let subjects = [];
+  for (let i = INT; i > ZERO; i--) {
+    const obj = {
+      isPublished,
+      absolutePath: i,
+    };
+    obj[fieldName] = i;
+    subjects.push(obj);
+  }
+  return subjects;
+}
+
+module.exports = {
+  getSubjects,
+};
 
 describe('Config perspective functions', () => {
   const ZERO = 0;
